shared/pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "shared"
version = "0.1.0"
description = "Shared utilities for Agensus Lambda functions"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.76",
    "httpx>=0.27.0",
    "langchain-text-splitters>=0.3.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "qdrant-client>=1.17.1",
    "voyageai>=0.3.7",
]

[tool.setuptools]
packages = ["shared"]

[tool.setuptools.package-dir]
shared = "."
